requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-multipart
boto3
MLflow
numpy
Pillow
scikit-image
uvicorn[standard]
websockets